2주차/requirements.txt
# YOLO 실험 환경 패키지 목록

# 핵심 딥러닝 프레임워크
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# 데이터 처리 및 분석
numpy>=1.21.0
pandas>=1.3.0
opencv-python>=4.5.0
Pillow>=8.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# 시각화
plotly>=5.0.0
wandb>=0.13.0  # 실험 추적 (선택사항)

# 데이터셋 다운로드 및 처리
requests>=2.25.0
tqdm>=4.62.0
pyyaml>=5.4.0

# 성능 최적화
onnx>=1.12.0
onnxruntime>=1.12.0
tensorrt>=8.4.0  # NVIDIA GPU 사용 시

# 유틸리티
scikit-learn>=1.0.0
scipy>=1.7.0
albumentations>=1.1.0  # 데이터 증강

# 개발 도구
jupyter>=1.0.0
ipykernel>=6.0.0
black>=22.0.0  # 코드 포맷팅
flake8>=4.0.0  # 코드 린팅

# 실험 관리
hydra-core>=1.2.0  # 설정 관리
omegaconf>=2.2.0

# 로깅 및 모니터링
tensorboard>=2.8.0
psutil>=5.8.0  # 시스템 모니터링

# 이미지 처리 추가 라이브러리
imageio>=2.9.0
scikit-image>=0.19.0

# 데이터셋 특화
pycocotools>=2.0.0  # COCO 데이터셋
lxml>=4.6.0  # XML 파싱

# GPU 모니터링
nvidia-ml-py3>=7.352.0  # NVIDIA GPU 모니터링

# 웹 인터페이스 (선택사항)
streamlit>=1.20.0  # 웹 대시보드
gradio>=3.0.0  # 웹 인터페이스

# 테스트
pytest>=6.2.0
pytest-cov>=3.0.0 